requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
boto3==1.34.154
botocore==1.34.154
build==1.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
fastapi==0.112.0
Flask==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pip-tools==7.4.1
pydantic==2.8.2
pydantic_core==2.20.1
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
s3transfer==0.10.2
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.5
Werkzeug==3.0.3
wheel==0.43.0
